requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
black==25.1.0
celery==5.4.0
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.1
cron-descriptor==1.4.5
cycler==0.12.1
distlib==0.3.9
Django==5.1.5
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.17.0
fonttools==4.56.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-tools==1.71.0
homebrew==0.0.0.1
identify==2.6.7
idna==3.10
kiwisolver==1.4.8
kombu==5.4.2
matplotlib==3.10.1
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pre_commit==4.1.0
prompt_toolkit==3.0.50
protobuf==5.29.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyparsing==3.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
seaborn==0.13.2
setuptools==80.1.0
six==1.17.0
SQLAlchemy==2.0.37
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.2
wcwidth==0.2.13
xmltodict==0.14.2
